+++ src/Ydb.Sdk/src/Metadata.cs
@@ -21,6 +21,4 @@ internal static class Metadata
     public const string GracefulShutdownHint = "session-close";
 
     internal static readonly string AdoNetClientInfo = $"ado-net/{YdbSdkVersion.Value}";
-    internal static readonly string TopicWriterClientInfo = $"topic-writer/{YdbSdkVersion.Value}";
-    internal static readonly string TopicReaderClientInfo = $"topic-reader/{YdbSdkVersion.Value}";
 }